Fix Online Store header and implementation
