+++ src/online_store_categories.c
@@ -7,7 +7,7 @@
 
 extern const struct TypeInfo gTypesInfo[NUMBER_OF_MON_TYPES];
 
-static const u8 sCategoryName_Balls[] = _("Pok\xE9 Balls");
+static const u8 sCategoryName_Balls[] = _("Poke Balls");
 static const u8 sCategoryName_Medicine[] = _("Medicine");
 static const u8 sCategoryName_TMTR[] = _("TMs/TRs");
 
@@ -24,20 +24,6 @@ static bool8 IsMedicine(u16 itemId)
 static bool8 IsTMTR(u16 itemId)
 {
     return gItemsInfo[itemId].pocket == POCKET_TM_HM;
-}
-
-static bool8 IsTypeFormItem(u16 itemId)
-{
-    const u8 *name = ItemId_GetName(itemId);
-    u8 i;
-
-    for (i = 0; i < NUMBER_OF_MON_TYPES; i++)
-    {
-        if (StringContains(name, gTypesInfo[i].name))
-            return TRUE;
-    }
-
-    return FALSE;
 }
 
 const struct OnlineStoreCategory gOnlineStoreCategories[] =
